rename and add build script
