+++ native/hub/src/lib.rs
@@ -18,7 +18,7 @@ mod sfx;
 mod stat;
 mod system;
 mod utils;
-mod macos_bridge;
+mod apple_bridge;
 
 use std::sync::Arc;
